Move stripe utilities to jwst.lib
