+++ jwst/lib/pipe_utils.py
@@ -150,204 +150,3 @@ def match_nans_and_flags(input_model):
         if data.shape != data_shape:
             continue
         data[is_invalid] = np.nan
-
-
-def generate_substripe_ranges(sci_model, subarray_ranges=False):
-    """
-    Determine the bounds of each substripe based on the input model multistripe metadata.
-
-    Parameters
-    ----------
-    sci_model : `~stdatamodels.jwst.datamodels.JwstDataModel`
-        The input datamodel with multistripe params defined.
-    subarray_ranges : bool
-        If true, dict containing ranges in subarray frame will also be provided.
-
-    Returns
-    -------
-    dict or tuple(dict)
-        Dictionary with keys as int counter, values as ranges of array index in slowaxis
-        corresponding to stripe shapes. If subarray_ranges was True, a second dictionary
-        will be provided with equivalent ranges in the subarray frame.
-    """
-    nreads1 = sci_model.meta.subarray.multistripe_reads1
-    nskips1 = sci_model.meta.subarray.multistripe_skips1
-    nreads2 = sci_model.meta.subarray.multistripe_reads2
-    nskips2 = sci_model.meta.subarray.multistripe_skips2
-    repeat_stripe = sci_model.meta.subarray.repeat_stripe
-    interleave_reads1 = sci_model.meta.subarray.interleave_reads1
-    slowaxis = sci_model.meta.subarray.slowaxis
-    if np.abs(slowaxis) > 1:
-        slowsize = sci_model.meta.subarray.ysize
-    else:
-        slowsize = sci_model.meta.subarray.xsize
-
-    sub_ranges = {}
-    ranges = {}
-    range_counter = 0
-    # Track the read position in the full frame with linecount, and number of lines
-    # read into subarray with sub_lines
-    linecount = 0
-    sub_lines = 0
-
-    # Start at 0, make nreads1 row reads
-    linecount += nreads1
-    sub_lines += nreads1
-    # Now skip nskips1
-    linecount += nskips1
-    # Nreads2
-    ranges[range_counter] = [linecount, linecount + nreads2]
-    sub_ranges[range_counter] = [sub_lines, sub_lines + nreads2]
-    range_counter += 1
-    linecount += nreads2
-    sub_lines += nreads2
-
-    # Now, while the output size is less than the science array size:
-    # 1a. If repeat_stripe, reset linecount (HEAD) to initial position
-    #     after every nreads2.
-    # 1b. Else, do nskips2 followed by nreads2 until subarray complete.
-    # 2.  Following 1a., repeat sequence of nreads1, skips*, nreads2
-    #     until complete. For skips*:
-    # 3a. If interleave_reads1, value of skips increments by nreads2 +
-    #     nskips2 for each stripe read.
-    # 3b. If not interleave, each loop after linecount reset is simply
-    #     nreads1 + nskips1 + nreads2.
-    interleave_skips = nskips1
-    if nreads2 <= 0:
-        raise ValueError(
-            "Invalid value for multistripe_reads2 - "
-            "cutout for reference file could not be "
-            "generated!"
-        )
-    while sub_lines < slowsize:
-        # If repeat_stripe, add interleaved rows to output and increment sub_lines
-        if repeat_stripe > 0:
-            linecount = 0
-            linecount += nreads1
-            sub_lines += nreads1
-            if interleave_reads1:
-                interleave_skips += nskips2 + nreads2
-                linecount += interleave_skips
-            else:
-                linecount += nskips1
-        else:
-            linecount += nskips2
-        ranges[range_counter] = [linecount, linecount + nreads2]
-        sub_ranges[range_counter] = [sub_lines, sub_lines + nreads2]
-        range_counter += 1
-        linecount += nreads2
-        sub_lines += nreads2
-
-    if sub_lines != slowsize:
-        raise ValueError(
-            "Stripe readout resulted in mismatched reference array shape "
-            "with respect to science array!"
-        )
-
-    if slowaxis < 0:
-        for rge in ranges.keys():
-            for i, row in enumerate(ranges[rge]):
-                ranges[rge][i] = 2048 - row
-                sub_ranges[rge][i] = 2048 - row
-
-    if subarray_ranges:
-        return ranges, sub_ranges
-    else:
-        return ranges
-
-
-def generate_superstripe_ranges(sci_model):
-    """
-    Return a dict of slowaxis ranges read into stripes.
-
-    Given an input model with multistripe parameters in its metadata,
-    return the slowaxis pixel ranges corresponding
-    to the nreads2 positions read into each stripe of a set
-    of superstripes. Dictionary keys are 0-indexed stripe
-    labels.
-
-    Parameters
-    ----------
-    sci_model : `~stdatamodels.jwst.datamodels.JwstDataModel`
-        The input datamodel with multistripe params defined.
-
-    Returns
-    -------
-    dict
-        Dictionary with keys as int counter, values as ranges of array index in slowaxis
-        corresponding to stripe shapes.
-    """
-    nreads1 = sci_model.meta.subarray.multistripe_reads1
-    nskips1 = sci_model.meta.subarray.multistripe_skips1
-    nreads2 = sci_model.meta.subarray.multistripe_reads2
-    nskips2 = sci_model.meta.subarray.multistripe_skips2
-    repeat_stripe = sci_model.meta.subarray.repeat_stripe
-    interleave_reads1 = sci_model.meta.subarray.interleave_reads1
-    superstripe_step = sci_model.meta.subarray.superstripe_step
-    num_superstripe = sci_model.meta.subarray.num_superstripe
-    xsize_sci = sci_model.meta.subarray.xsize
-    ysize_sci = sci_model.meta.subarray.ysize
-    fastaxis = sci_model.meta.subarray.fastaxis
-
-    if np.abs(fastaxis) == 1:
-        slow_size = ysize_sci
-    else:
-        slow_size = xsize_sci
-
-    ranges = {}
-    range_counter = 0
-
-    for stripe in range(num_superstripe):
-        # Track the read position in the full frame with linecount, and number of lines
-        # read into subarray with sub_lines
-        linecount = 0
-        sub_lines = 0
-
-        # Start at 0, make nreads1 row reads
-        linecount += nreads1
-        sub_lines += nreads1
-        # Now skip nskips1 + superstripe_step * stripe
-        linecount += nskips1 + superstripe_step * stripe
-        # Nreads2
-        ranges[range_counter] = [(linecount, linecount + nreads2)]
-        linecount += nreads2
-        sub_lines += nreads2
-
-        # Now, while the output size is less than the science array size:
-        # 1a. If repeat_stripe, reset linecount (HEAD) to initial position
-        #     after every nreads2.
-        # 1b. Else, do nskips2 followed by nreads2 until subarray complete.
-        # 2.  Following 1a., repeat sequence of nreads1, skips*, nreads2
-        #     until complete. For skips*:
-        # 3a. If interleave_reads1, value of skips increments by nreads2 +
-        #     nskips2 for each stripe read.
-        # 3b. If not interleave, each loop after linecount reset is simply
-        #     nreads1 + nskips1 + nreads2.
-        interleave_skips = nskips1
-
-        while sub_lines < slow_size:
-            # If repeat_stripe, add interleaved rows to output and increment sub_lines
-            if repeat_stripe > 0:
-                linecount = 0
-
-                linecount += nreads1
-                sub_lines += nreads1
-
-                if interleave_reads1:
-                    interleave_skips += nskips2 + nreads2
-                    linecount += interleave_skips
-                else:
-                    linecount += nskips1
-            else:
-                linecount += nskips2
-            ranges[range_counter].append((linecount, linecount + nreads2))
-            linecount += nreads2
-            sub_lines += nreads2
-        range_counter += 1
-
-        if sub_lines != slow_size:
-            raise ValueError(
-                "Stripe readout resulted in mismatched reference array shape "
-                "with respect to science array!"
-            )
-    return ranges
